chore(*): organize classes more
